+++ ontimize/config/o-modules.ts
@@ -1,81 +1,76 @@
+import { HttpClient, HttpClientModule } from '@angular/common/http';
+import { APP_INITIALIZER, Injector, NgModule } from '@angular/core';
 import { HttpModule } from '@angular/http';
-import { HttpClientModule, HttpClient } from '@angular/common/http';
-import { TranslateModule, TranslateLoader, TranslateParser } from '@ngx-translate/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { Injector, NgModule, APP_INITIALIZER } from '@angular/core';
+import { DndModule } from '@churchs19/ng2-dnd';
+import { TranslateLoader, TranslateModule, TranslateParser } from '@ngx-translate/core';
 import { NgxMaterialTimepickerModule } from 'ngx-material-timepicker';
 
-import { appInitializerFactory } from './o-providers';
-import { OTranslateService, OTranslateHttpLoader } from '../services';
-import { APP_CONFIG, AppConfig } from '../config/app-config';
-
 import {
-  OBarMenuModule,
   OBarMenuGroupModule,
   OBarMenuItemModule,
-  OLocaleBarMenuItemModule,
-  OBarMenuSeparatorModule,
+  OBarMenuModule,
   OBarMenuNestedModule,
+  OBarMenuSeparatorModule,
   OBreadcrumbModule,
   OButtonModule,
   OButtonToggleModule,
+  OCardMenuItemModule,
   OCheckboxModule,
-  OComboModule,
+  OColumnCollapsibleModule,
   OColumnModule,
+  OComboModule,
   OContextMenuModule,
-  ORowModule,
-  ODialogModule,
-  OSnackBarModule,
-  OFilterBuilderModule,
-  OFormModule,
-  OFormContainerModule,
-  OImageModule,
   OCurrencyInputModule,
   ODateInputModule,
+  ODialogModule,
   OEmailInputModule,
+  OErrorModule,
   OFileInputModule,
+  OFilterBuilderModule,
+  OFormContainerModule,
+  OFormModule,
+  OGridModule,
+  OHourInputModule,
   OHTMLInputModule,
+  OImageModule,
   OIntegerInputModule,
+  OLanguageSelectorModule,
+  OListItemAvatarModule,
+  OListItemCardImageModule,
+  OListItemCardModule,
+  OListItemModule,
+  OListItemTextModule,
+  OListModule,
   OListPickerModule,
+  OLocaleBarMenuItemModule,
   ONIFInputModule,
   OPasswordInputModule,
   OPercentInputModule,
-  ORealInputModule,
-  OTextInputModule,
-  OTextareaInputModule,
-  OValidatorModule,
-  OErrorModule,
-  OListModule,
-  OListItemModule,
-  OListItemAvatarModule,
-  OListItemTextModule,
-  OListItemCardModule,
-  OListItemCardImageModule,
-  OSearchInputModule,
-  OTableModule,
-  OUserInfoModule,
-  OLanguageSelectorModule,
-  OCardMenuItemModule,
-  OHourInputModule,
-  OTimeInputModule,
-  OGridModule,
   ORadioModule,
+  ORealInputModule,
+  ORowCollapsibleModule,
+  ORowModule,
+  OSearchInputModule,
+  OSliderModule,
   OSlideToggleModule,
-  OSliderModule
+  OSnackBarModule,
+  OTableModule,
+  OTextareaInputModule,
+  OTextInputModule,
+  OTimeInputModule,
+  OUserInfoModule,
+  OValidatorModule,
 } from '../components';
-
-import {
-  OAppLayoutModule,
-  OFormLayoutManagerModule,
-  OCardMenuLayoutModule
-} from '../layouts';
-
-import { OSharedModule } from '../shared';
-import { OTranslateParser } from '../services/translate/o-translate.parser';
-import { Util } from '../utils';
-import { DndModule } from '@churchs19/ng2-dnd';
+import { APP_CONFIG, AppConfig } from '../config/app-config';
+import { OAppLayoutModule, OCardMenuLayoutModule, OFormLayoutManagerModule } from '../layouts';
+import { OTranslateHttpLoader, OTranslateService } from '../services';
 import { OPermissionsModule } from '../services/permissions/o-permissions.module';
+import { OTranslateParser } from '../services/translate/o-translate.parser';
+import { OSharedModule } from '../shared';
+import { Util } from '../utils';
+import { appInitializerFactory } from './o-providers';
 
 export const INTERNAL_ONTIMIZE_MODULES_EXPORTED: any = [
   // Standard modules
@@ -96,8 +91,10 @@ export const INTERNAL_ONTIMIZE_MODULES_EXPORTED: any = [
   OCheckboxModule,
   OComboModule,
   OColumnModule,
+  OColumnCollapsibleModule,
   OContextMenuModule,
   ORowModule,
+  ORowCollapsibleModule,
   ODialogModule,
   OSnackBarModule,
   OFilterBuilderModule,
@@ -194,8 +191,10 @@ export const INTERNAL_ONTIMIZE_MODULES: any = [
   OCheckboxModule,
   OComboModule,
   OColumnModule,
+  OColumnCollapsibleModule,
   OContextMenuModule,
   ORowModule,
+  ORowCollapsibleModule,
   ODialogModule,
   OSnackBarModule,
   OFilterBuilderModule,